+++ R/akin-package.R
@@ -1,12 +1,11 @@
-#'
-#' @name akin-package
 #' @title Functional Utilities For Data Processing
+#' @name akin-package
 #'
 #' @description
 #' The intent of this package is utilitarian which explains the relatively large number of informative messages
 #' displayed by most functions. Designed for large and very large data files, the package employs data.table,
-#' list environments, sparse matrices, iterators and background processing in few places. Nevertheless, Users could
-#' consider thread optimization, memoisation or parallel processing, as these techniques were outside the scope.
+#' list environments, sparse matrices, iterators, background and parallel processing in few places. Nevertheless,
+#' Users could consider thread optimization or memoisation, as these techniques were outside the scope.
 #'
 #' The package covers several areas of data processing: subset reading and writing of large data files, data tiling
 #' (horizontal and vertical splitting) - suited for data conversion operations with local as well as global hold,
@@ -16,11 +15,10 @@
 #' covering various representations of protein mass spectrometry data with no pretense of exhaustiveness.
 #'
 #' Comments and suggestions regarding tricky situations are provided however brief. Examples should be run individually
-#' in R console.
+#' in R console. When possible, suppression of messages should be avoided.
 #'
-#' @docType package
 #' @keywords internal
 #'
-
  "_PACKAGE"
  NULL
+