+++ NEWS.md
@@ -34,3 +34,16 @@ Added function 'common' which identifies common substrings inside a pair of chai
 
 * Added the experimental function "fcommon", a fast version of "common".
 
+# akin 0.3.1
+
+* Expanded function "fcommon" capability to parallel processing for substrings longer than 20 characters.
+  A local parallel plan is set when formal argument "strategy" receives a value. Pending on this, formal
+  argment "workers" may also receive a value. Without a parallel plan, the function will output the long
+  substrings as they are, along with a message. The output is no longer a closure but a character vector.
+  
+* Function "common" remains available.
+ 
+* Modified the Manual accordingly.
+
+* Added package "future" to Suggests and package "future.apply" to Depends in DESCRIPTION.
+